+++ source/d3d11/d3d11_impl_device.cpp
@@ -1194,7 +1194,7 @@ bool reshade::d3d11::device_impl::get_query_pool_results(api::query_pool pool, u
 	com_ptr<ID3D11DeviceContext> immediate_context;
 	_orig->GetImmediateContext(&immediate_context);
 
-	for (uint32_t i = 0; i < count; ++i)
+	for (size_t i = 0; i < count; ++i)
 	{
 		if (FAILED(immediate_context->GetData(impl->queries[first + i].get(), static_cast<uint8_t *>(results) + i * stride, stride, D3D11_ASYNC_GETDATA_DONOTFLUSH)))
 			return false;
